requirements.txt
streamlit
kagglehub
pandas
sentence-transformers
python-dotenv